+++ src/features/harborView/HarborViewPage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 
 import styles from './harborViewPage.module.scss';
-import PageTitle from '../../../common/pageTitle/PageTitle';
+import PageTitle from '../../common/pageTitle/PageTitle';
 
 const HarborViewPage: React.SFC = ({ children }) => {
   const { t } = useTranslation();
